requirements.txt
langchain
langchain-community
langchain-groq
langchain-core
yfinance
python-dotenv
pandas
streamlit
duckduckgo-search
